Fixing progress-bar behavior
