+++ src/styles/components/popup.less
@@ -1,5 +1,5 @@
 .popup {
-  z-index: 999;
+  z-index: 9;
 
   // preventing animation on load
   &.loaded {
