+++ src/styles/screens/form-screen/form-screen.less
@@ -40,6 +40,10 @@
     max-height: none;
     justify-content: flex-start;
 
+    .confetti-container {
+      position: absolute;
+    }
+
     .form-top-section:not(:empty) {
       text-align: left;
       display: flex;
